+++ tests/Datasets/Generators/CircleTest.php
@@ -3,7 +3,7 @@
 namespace Rubix\ML\Tests\Datasets\Generators;
 
 use Rubix\ML\Datasets\Dataset;
-use Rubix\ML\Datasets\Unlabeled;
+use Rubix\ML\Datasets\Labeled;
 use Rubix\ML\Datasets\Generators\Circle;
 use Rubix\ML\Datasets\Generators\Generator;
 use PHPUnit\Framework\TestCase;
@@ -27,7 +27,7 @@ class CircleTest extends TestCase
     {
         $dataset = $this->generator->generate(30);
 
-        $this->assertInstanceOf(Unlabeled::class, $dataset);
+        $this->assertInstanceOf(Labeled::class, $dataset);
         $this->assertInstanceOf(Dataset::class, $dataset);
 
         $this->assertCount(30, $dataset);